fixed wav2vec half conversion issue

+++ fairseq/models/wav2vec/wav2vec2.py
@@ -426,7 +426,7 @@ class Wav2Vec2Model(BaseFairseqModel):
                 min_space=self.mask_min_space,
             )
             mask_indices = torch.from_numpy(mask_indices).to(x.device)
-            x[mask_indices] = self.mask_emb
+            x[mask_indices] = self.mask_emb.to(x.dtype)
         else:
             mask_indices = None
